Make delays in eq blocks a little longer
Making the delays in the eq blocks use behavioral HDL did give the tools more
flexibility in how to handle the delays.  Interestingly, the tools chose to
push the two cycles of latency into the DSP48 element (the inputs can have
internal latencies of 0, 1, or 2) which effectively created a single cycle path
from the BRAM to the DSP48 (i.e. the multiplier).  Without any pipelining delay
along the route timing still failed in a few places.

This change makes the delays have a latency of 4 cycles.  Even if two cycles
are pushed into the DSP48, the tools will still have two cycles of latency to
spread out along the path (if they choose to do so).