Add mode port to transpose_core
The mode port selects between four different transpose mappings.  The mappings
allow a subset of (or all) F engines to drive a subset of (or all) X engines.
When driving a subset of X engines from a subset of F engines, channels from
the subset of F engines that would normally be processed by "missing" X engines
are masqueraded as different channels from the "missing" F engines.

The four modes correspond to using 8, 4, 2, or 1 F engine to drive 32, 16, 8,
or 4 X engines, resp.  The modes are enumerated as:

mode == 0 is the 8 F engine mode.  This is the complete system, no masquerading is
needed.  Use this for normal operation.

mode == 1 is the 4 F engine mode.  This mode uses 4 F engines to drive 16 X
engines.  In this mode, half of the band masquerades as the other half of the
band for the 4 "missing" F engines.

mode == 2 is the 2 F engine mode.  This mode uses 2 F engines to drive 8 X
engines.  In this mode, three quarters of the band masqueade as the other
quarter of the band for the 6 "missing" F engines.

mode == 3 is the 1 F engine mode.  This mode uses 1 F engine to drive 4 X
engines.  In this mode, the seven eigths of the band masquerade as the other
eigth of the band for the 7 "missing" F engines.

All modes other than mode == 0 are for testing/diagnostic purposes only; they
are not intended for use in the deployed system.