Add snap block
Add snapshot capability to roach2_fengine.  There is one snapshot block that is
4k deep by 32 bits wide.  The input to the snap block is multiplexed from
various points in the signal path based on the value of a selector software
register.  The snapshot block is named "snap/0" and the selector register is
named "snap/0_sel".  The "0" was chosen in anticipation of possible additional
snapshot blocks in the future (e.g. "1").  No support for synchronizing the
snapshot with is provided; it snaps when the software tells it to so the
snapshot time is roughly known within the mushiness of the software timing.

The snapshot mux selects from four possible points in the signal path:

snap_sel_0     snap_0 input
               and 32-bit word format
----------     -----------------------
    0          Output of the input mux
               {in3, in2, in1, in0} (four Fix_8_7 values)

    1          Output of the PFB preshifter
               {in0} (Fix_18_17 sign enxtended to 32 bits)

    2          Output of the FFT block
               {re(in0)} (Fix_18_17 sign enxtended to 32 bits)

    3          Output of the EQ block
               {in3, in2, in1, in0} (Four 4bit + 4bit complex values)

Note that the output of the FFT block cycles through 1024 samples of
input 0 followed by 1024 samples of input 16.

Note that the output of the EQ block cycles through 1024 samples of
inputs 0,1,2,3 followed by 1024 samples of inputs 16,17,18,19.