Fix input level RMS logic
The Fix_8_7 input went through a Convert block to become a Fix_8_0 data type.
Unfortunately, this should have been a Reinterpret block.  Using a convert
block effectively changes every sample to 0 or -1.  This means that the sum is
always the negative of the sum of squares.

This commit replaces the 32 Convert blocks (one per input) with Reinterpret
blocks to force the binary point of the values to 0 without changing the bits.